+++ src/sys/select.rs
@@ -10,6 +10,8 @@ pub use libc::FD_SETSIZE;
 
 // FIXME: Change to repr(transparent) once it's stable
 #[repr(C)]
+#[derive(Clone, Copy)]
+#[allow(missing_debug_implementations)]
 pub struct FdSet(libc::fd_set);
 
 impl FdSet {
@@ -56,7 +58,7 @@ impl FdSet {
     ///
     /// [`select`]: fn.select.html
     pub fn highest(&mut self) -> Option<RawFd> {
-        for i in (0..libc::FD_SETSIZE).rev() {
+        for i in (0..FD_SETSIZE).rev() {
             let i = i as RawFd;
             if unsafe { libc::FD_ISSET(i, self as *mut _ as *mut libc::fd_set) } {
                 return Some(i)